requirements.txt
streamlit>=1.42
anthropic>=0.52
Pillow>=10.0
markdown>=3.5
